Tests in app and add confirm alert to remove an event

+++ src/components/ui/DeleteEventFab.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { useDispatch } from 'react-redux'
+import Swal from 'sweetalert2';
 import { eventStartDeleteAction } from '../../actions/events';
 
 export const DeleteEventFab = () => {
@@ -7,7 +8,25 @@ export const DeleteEventFab = () => {
     const dispatch = useDispatch();
 
     const handleDelete = () => {
-        dispatch(eventStartDeleteAction());
+        Swal.fire({
+            title: '¿Está seguro?',
+            text: "No podrá revertir esto!!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, elimimar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(eventStartDeleteAction());
+                Swal.fire(
+                    'Eliminado!!',
+                    'Tu evento ha sido borrado',
+                    'success'
+                )
+            }
+        })
     }
 
     return (
